Use 32 bit to make it work on 32 bit

+++ interfaces/opt.go
@@ -30,5 +30,5 @@ type ExtraOptionsBoostConnectionGracefulShutdownRequestValue struct {
 const ExtraOptionsBoostConnectionShouldNotRedial = "ExtraOptionsBoostConnectionShouldNotRedial"
 
 type ExtraOptionsBoostConnectionShouldNotRedialValue struct {
-	ShouldNotReDial int64
+	ShouldNotReDial int32
 }
